ConstantPool Types
ConstantPool is composed by some primitive types should be from the ConstantPoolType. String, Integer, MethodRef, NameAndType, etc.. Are implemented in this commit

+++ src/main/java/com/allemas/classfile/constantpool/Double.java
@@ -11,7 +11,7 @@ public class Double extends ConstantPoolInfo {
         this.value = value;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
